EventListener: Fix bug where avaialble events would not show
One possible cause is that the rendered list being saved much sooner
than the saved notice gets saved. Another cause is that when the saved
notice times out, it gets a new notice. Either of these events caused
an old rendered list to be used, which caused events to not be shown.

The rendered list is for when the notice is saved and to carry over
the current status of rendered events upon a page reload so that the
same events do not get rerendered when using the check event controls.