IndesxedAutocomplete: Fix the enter key being propagated at selection
The issue is that the rebind happens late in the process, so the
autocomplete handler is at the end of the queue instead of at the
beginning, so all other handlers, to include the hotkey handlers,
were getting processed first. This commit moves the autocomplete
handler to the beginning of the queue.