IndexedAutocomplete: Fix tag categorey and operator prefix autocomplete
One issue is that the prefixes weren't being joined with the pipe operator.
Another is that the regex is now inherently broken, since it will replace
the category or the operator, but not both. Since this match will always
be non-whitespace, that text is removed and the calculated prefix is
now added.