New TISAS: Fix issue where tweets would jump when clicking links
This seems to only happen on Chrome, and searching the internet there
was at least one other case of someone complaining about something
similar with Chrome. It was discovered that it was the CSS border
property which was causing the issue. So instead of using borders,
4 line segments were created with absolute positioning off of the
Tweet itself. This seems to have mostly fixed the issue.