Debug server: Initial commit
Provides Python scripts which allows one to quickly change the library
CSS being tested, so that one doesn't have to push updates to Github
and then update the commit hashes in all of the scripts being tested.

1. Modify the CSS in the files under the CSS directory
2. Execute the `css_to_json.py` script
3. Excecute the `file_server.py` script

The file server will only release the command prompt by pressing
"Ctrl+C", so it may be necessary to open up 2 command windows.