IndexedAutocomplete: Update CSS values for expandable section
Remove page specificity, since the number of pages has grown larger
with the addition of upload media assets. Also remove obsolete CSS,
since Danbooru no longer uses wide tag groups with the removal of the
artist section. Because of this, the base width was increased to give
the columns more room to expand before they start to wrap.