New TISAS: Fix issue of duplicate image links presented on some videos
Mentioned on issue #5. This seems to only happen on Chrome. It was
seen that it was marking two different elements with duplicate image
url data attributes. This only happened when auto-unhide was being used.

It does a check against all existing image URLs to ensure that that
the image URL being added is not already seen on the tweet. Since this
only affects the tweet page, this check was limited to that page, to
avoid potential issues this may cause on the other stream views.