PostModeMenu: Prevent middle click from firing tag mode events
The DragSelect addon was including the middle button as one allowed
to initiate drag/select events, which was giving the appearance that
clicking with the middle key would sometimes cause the tag mode event
to fire with no consistency. However, the slightest of motion will cause
the event to be a drag event and not a click event. However, to the
user, it will appear as if the the click did not work properly, when
really the system did not register it as a click event at all.

Fixes #9.