New TISAS: Break the media markup loop early on non-media tweets
This was causing some tweets to be incorrectly marked as image tweets
since that is the default case. Even when some of the subelements got
correctly marked as non-media, it would still continue to process the
remaining elements, which often fell under that default case and were
marked as image media.