New TISAS: Move several classes to using attributes instead
This was causing conflicts with extensions like uBlock which rely
heavily on precise class usage. For example, several of the selectors
specified that one-and-only-one class could exist on an element.

The similarity classes weren't in conflict with anything, but using
an attribute instead matches the highlights and plus its less complex.