New TISAS: Create a placeholder function for getting multiple tweets
This might be used in the future as a backup to getting the API data
for tweets. There is no GraphQL equivalent, so it'd only be for those
that can use the 1.1 API with their browser/extension/accounts.

Also export the API functions during debugging so they can be tested.