EventListener: Prevent non-subscribed types from modifying the list
This was causing issues when an event type was autosubscribed but
not enabled. On the post SHOW page, the multi-link menu renders
every event type, even if not enabled, meaning that the autosubscribe
event would activate, but since the event itself was not enabled, the
get list returned an empty set. This effectively was causing whatever
post ID that was being added to overwrite the entire list.