New TISAS: Add extra debug messages for normalizing webp links
It'd be better to make a separate API call for these in the download
image function rather than having to make unnecessary HEAD calls every
single time a webp image is found. This will also allow the chain of
functions that flow from this to be synchronous again. So these debug
messages will be used to find webp links out in the wild.