DisplayPostInfo: Possible fix for tooltip setting modification
It was reported that the settings don't work on the modqueue page,
which has been observed on the post SHOW page due to the presence
of user tooltips which override the Tippy instance on the body
variable. However, post tooltips saves the instance as a module
variable, which works on all of the pages I was able to test out.