New TISAS: Revise Twitter network functions
It was discovered that the 1.1 API doesn't always return the full array
of image/video data even when a tweet is returned, so those endpoints
were scrapped. TweetResultsByRestIds was added to the available GraphQL
functions. With two of each for single/multi-tweets, a primary/backup
scheme was incorporated into the tweets database functions.