New TISAS: Fix get of multiple tweets data
It was discovered that the same tweets were being queried more than
once on the media timeline, so a class was utilized to mark tweets
which have already been handled.

However, the memoize key wasn't really helping either, so the GetTweetsData
function was revised to create a memoize status for each individual tweet.