New TISAS: Make usage of the media API endpoint a user setting
This is because it needs to be kept up-to-date, and it consumes a
lot more tweets viewed than the other endpoint. This makes it so that
there is no longer a backup though.

The singular tweet endpoints don't have this issue, so it is being kept
with the current backup scheme, with the more common TweetDetail as
the primary endpoint.