New TISAS: Update the colors for dark mode
This was neglected in the original push. This required adding a few
markers for CSS selectors. While at it, a few items were renamed to
make it clear what they are. Additionally, several other CSS fixes were
done, including fixing the rgb/rgba of the rendered colors. The language
in some of the helps were updated, since the border color can now vary
depending on the background color mode.