New TISAS: Update CSS
Twitter changed the CSS a while ago, to include the font-size, which
messed with all of the sizes and alignment. So, every container where
items are set have their font-size set at 15px, which will keep all of
the em values at a constant. The side menu and settings menu already
had their font-sizes set at 14px, and won't be changed since that would
require updating all of the em values as well.

At the same tie, the tweet menu was simplified, which also helps with
keeping items aligned or centered.