EventListener: Mark forum topic as read when opening a forum post
Even though the original request gets sent with the .html extension,
the redirect does not, and so the type isn't specifically set as HTML,
even though that is what is returned. Specifying the accept as only
text/html instead of */* will cause the forum topics controller to
mark the forum topics as read.