Library-danbooru: Add/update several functions
Update post update function to avoid using the JS route which causes
issues with blacklists as well as other issue. Added several helper
functions, which can be used for other update-like functions. The
`pending_updates` variable required a getter and setter as well as
a reset, since the value is set to 0 by the always function, but it's
the success callback function which uses it. This was causing the
plural variable to be wrong depending on the function call order.